A small change in the list title and review pt_BR translation
When "no task list" appears listing the lists below the title,
the negation doesn't make much sense and yet I removed that
negation.

Along with this change, the Brazilian Portuguese translation
file was revised, the necessary changes were made, such as
adding credit to Guilherme Berbet and the update was made
following some lines of the file.